Removes incorrect DST and timezone adjustment
Unix timestamps, aka POSIX time, is always the number of seconds which
have lapsed since 1970/01/01 in UTC. The timezone adjustment and DST
adjustment should be done at client end when rendering the time to show
end-user, not on the server end.  Having this information modified here
means the client must know where the server is located and makes
further adjustments in order to render correctly.